Move addons package from cmd/juju to agent.
To faciliate adding an Introspection worker to the UnitAgents created in
the deployer worker.
