+++ agent/addons/addons.go
