+++ agent/addons/package_test.go
