+++ cmd/jujud/introspect/introspect.go
@@ -19,9 +19,9 @@ import (
 	"github.com/juju/names/v4"
 	"github.com/kr/pretty"
 
+	"github.com/juju/juju/agent/addons"
 	apiagent "github.com/juju/juju/api/agent"
 	jujucmd "github.com/juju/juju/cmd"
-	"github.com/juju/juju/cmd/jujud/agent/addons"
 	"github.com/juju/juju/cmd/jujud/agent/config"
 )
